# docs/swagger/openapi-converted.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "API for managing community watch shifts, bookings, and reports",
        "title": "Community Watch Shift Scheduler API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/admin/emergency-contacts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all emergency contacts for admin management",
                "tags": [
                    "admin-emergency-contacts"
                ],
                "summary": "Admin: Get emergency contacts",
                "responses": {
                    "200": {
                        "description": "List of emergency contacts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/api.EmergencyContactResponse"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - admin access required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new emergency contact",
                "tags": [
                    "admin-emergency-contacts"
                ],
                "summary": "Admin: Create emergency contact",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.CreateEmergencyContactRequest"
                            }
                        }
                    },
                    "description": "Emergency contact data",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created emergency contact",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.EmergencyContactResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - admin access required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/admin/emergency-contacts/{id}/default": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sets a specific emergency contact as the default",
                "tags": [
                    "admin-emergency-contacts"
                ],
                "summary": "Admin: Set default emergency contact",
                "parameters": [
                    {
                        "description": "Emergency Contact ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Default emergency contact updated successfully"
                    },
                    "400": {
                        "description": "Invalid contact ID",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - admin access required",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Emergency contact not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/emergency-contacts": {
            "get": {
                "description": "Returns all active emergency contacts for public use",
                "tags": [
                    "emergency-contacts"
                ],
                "summary": "Get emergency contacts",
                "responses": {
                    "200": {
                        "description": "List of emergency contacts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/api.EmergencyContactResponse"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/emergency-contacts/default": {
            "get": {
                "description": "Returns the default emergency contact (usually RUSA)",
                "tags": [
                    "emergency-contacts"
                ],
                "summary": "Get default emergency contact",
                "responses": {
                    "200": {
                        "description": "Default emergency contact",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.EmergencyContactResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No default emergency contact found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "http://localhost:5888"
        }
    ],
    "components": {
        "schemas": {
...
            "api.CreateEmergencyContactRequest": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "display_order": {
                        "type": "integer"
                    },
                    "is_default": {
                        "type": "boolean"
                    },
                    "name": {
                        "type": "string"
                    },
                    "number": {
                        "type": "string"
                    }
                }
            },
...
            "api.EmergencyContactResponse": {
                "type": "object",
                "properties": {
                    "description": {
                        "type": "string"
                    },
                    "display_order": {
                        "type": "integer"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "is_default": {
                        "type": "boolean"
                    },
                    "name": {
                        "type": "string"
                    },
                    "number": {
                        "type": "string"
                    }
                }
            },
...
            "api.ErrorResponse": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}